# Multithreaded wallet syncing