--- a/src/Utilities/ThreadSafePriorityQueue.h
+++ b/src/Utilities/ThreadSafePriorityQueue.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2019, The TurtleCoin Developers
+// Copyright (c) 2019, The TurtleCoin Developers
 // 
 // Please see the included LICENSE file for more information.
 
@@ -6,47 +6,43 @@
 
 #include <condition_variable>
 
-#include <deque>
+#include <queue>
 
 #include <mutex>
 
-#include <iostream>
-
-template<typename T>
-class ThreadSafeDeque
+template<typename T, typename Comparison = std::less<T>>
+class ThreadSafePriorityQueue
 {
     public:
-        ThreadSafeDeque() :
+        ThreadSafePriorityQueue() :
             m_shouldStop(false)
         {
         }
 
-        ThreadSafeDeque(bool startStopped) :
+        ThreadSafePriorityQueue(bool startStopped) :
             m_shouldStop(startStopped)
         {
         }
 
         /* Move constructor */
-        ThreadSafeDeque(ThreadSafeDeque && old)
+        ThreadSafePriorityQueue(ThreadSafePriorityQueue && old)
         {
             *this = std::move(old);
         }
 
-        ThreadSafeDeque& operator=(ThreadSafeDeque && old)
+        ThreadSafePriorityQueue& operator=(ThreadSafePriorityQueue && old)
         {
             /* Stop anything running */
             stop();
 
-            m_deque = std::move(old.m_deque);
+            m_priorityQueue = std::move(old.m_priorityQueue);
 
             return *this;
         }
 
-        /* Add the items to the end of the queue. They are added in the order
-           of the iterators passed in, so once the operation has completed,
-           the item pointed to by end will be at the end of the queue. */
+        /* Add the items between [begin, end] to the queue */
         template<typename Iterator>
-        bool push_back_n(Iterator begin, Iterator end)
+        bool push_n(Iterator begin, Iterator end)
         {
             /* Aquire the lock */
             std::unique_lock<std::mutex> lock(m_mutex);
@@ -59,7 +55,7 @@ class ThreadSafeDeque
 
             for (auto it = begin; it < end; it++)
             {
-                m_deque.push_back(*it);
+                m_priorityQueue.push(*it);
             }
 
             /* Unlock the mutex before notifying, so it doesn't block after
@@ -73,7 +69,7 @@ class ThreadSafeDeque
         }
 
         /* Add an item to the end of the queue */
-        bool push_back(T item)
+        bool push(T item)
         {
             /* Aquire the lock */
             std::unique_lock<std::mutex> lock(m_mutex);
@@ -84,8 +80,7 @@ class ThreadSafeDeque
                 return false;
             }
 
-            /* Add the item to the front of the queue */
-            m_deque.push_back(item);
+            m_priorityQueue.push(item);
 
             /* Unlock the mutex before notifying, so it doesn't block after
                waking up */
@@ -98,7 +93,7 @@ class ThreadSafeDeque
         }
 
         /* Delete the front item from the queue */
-        void pop_front()
+        void pop()
         {
             /* Aquire the lock */
             std::unique_lock<std::mutex> lock(m_mutex);
@@ -111,13 +106,13 @@ class ThreadSafeDeque
                force a removal short of running it in a while loop.
                Instead, if we just force the queue to have to have data in,
                we can make sure a removal always suceeds. */
-            if (m_deque.empty())
+            if (m_priorityQueue.empty())
             {
                 throw std::runtime_error("Cannot remove from an empty queue!");
             }
 
             /* Remove the first item from the queue */
-            m_deque.pop_front();
+            m_priorityQueue.pop();
 
             /* Unlock the mutex before notifying, so it doesn't block after
                waking up */
@@ -126,8 +121,26 @@ class ThreadSafeDeque
             m_consumedData.notify_all();
         }
 
+        /* Delete the front item from the queue, without aquiring the lock */
+        void pop_unsafe()
+        {
+            m_priorityQueue.pop();
+        }
+
+        /* Get the top item, without aquiring the lock */
+        T top_unsafe()
+        {
+            return m_priorityQueue.top();
+        }
+
+        /* Determine if the container is empty, without aquiring the lock */
+        bool empty_unsafe()
+        {
+            return m_priorityQueue.empty();
+        }
+
         /* Removes numElements from the start of the queue */
-        void pop_front_n(size_t numElements)
+        void pop_n(size_t numElements)
         {
             /* Aquire the lock */
             std::unique_lock<std::mutex> lock(m_mutex);
@@ -140,12 +153,12 @@ class ThreadSafeDeque
                force a removal short of running it in a while loop.
                Instead, if we just force the queue to have to have data in,
                we can make sure a removal always suceeds. */
-            if (m_deque.empty())
+            if (m_priorityQueue.empty())
             {
                 throw std::runtime_error("Cannot remove from an empty queue!");
             }
 
-            if (m_deque.size() < numElements)
+            if (m_priorityQueue.size() < numElements)
             {
                 throw std::runtime_error("Cannot remove more elements than are stored!");
             }
@@ -153,7 +166,7 @@ class ThreadSafeDeque
             while (numElements > 0)
             {
                 /* Remove first item from queue */
-                m_deque.pop_front();
+                m_priorityQueue.pop_front();
                 numElements--;
             }
 
@@ -165,43 +178,15 @@ class ThreadSafeDeque
         }
 
         /* Take an item from the front of the queue, and do NOT remove it */
-        T front()
+        T top()
         {
             return getFirstItem(false);
         }
 
         /* Take and remove an item from the front of the queue */
-        T front_and_remove()
+        T top_and_remove()
         {
             return getFirstItem(true);
-        }
-
-        std::vector<T> front_n_and_remove(const size_t numElements)
-        {
-            /* Aquire the lock */
-            std::unique_lock<std::mutex> lock(m_mutex);
-
-            std::vector<T> results;
-
-            if (m_deque.size() == 0)
-            {
-                return results;
-            }
-
-            if (m_deque.size() <= numElements)
-            {
-                results.resize(m_deque.size());
-                std::copy(m_deque.begin(), m_deque.end(), results.begin());
-                m_deque.clear();
-            }
-            else
-            {
-                results.resize(numElements);
-                std::copy(m_deque.begin(), m_deque.begin() + numElements, results.begin());
-                m_deque.erase(m_deque.begin(), m_deque.begin() + numElements);
-            }
-
-            return results;
         }
 
         /* Stop the queue if something is waiting on it, so we don't block
@@ -231,56 +216,7 @@ class ThreadSafeDeque
                but hey) */
             std::unique_lock<std::mutex> lock(m_mutex);
 
-            return m_deque.size();
-        }
-
-        /* Takes n elements, if available, starting at the head of the queue.
-           Otherwise returns max available. Does not remove items from the queue. */
-        std::vector<T> front_n(const size_t numElements) const
-        {
-            std::cout << "Requested " << numElements << " blocks. Available: "
-                      << m_deque.size() << std::endl;
-
-            /* Aquire the lock */
-            std::unique_lock<std::mutex> lock(m_mutex);
-
-            std::vector<T> results;
-
-            if (m_deque.size() <= numElements)
-            {
-                results.resize(m_deque.size());
-                std::copy(m_deque.begin(), m_deque.end(), results.begin());
-            }
-            else
-            {
-                results.resize(numElements);
-                std::copy(m_deque.begin(), m_deque.begin() + numElements, results.begin());
-            }
-
-            return results;
-        }
-
-        /* Takes n elements, if available, starting at the tail of the queue.
-           Otherwise returns max available. Does not remove items from the queue. */
-        std::vector<T> back_n(const size_t numElements) const
-        {
-            /* Aquire the lock */
-            std::unique_lock<std::mutex> lock(m_mutex);
-
-            std::vector<T> results;
-
-            if (m_deque.size() <= numElements)
-            {
-                results.resize(m_deque.size());
-                std::copy(m_deque.rbegin(), m_deque.rend(), results.begin());
-            }
-            else
-            {
-                results.resize(numElements);
-                std::copy(m_deque.rbegin(), m_deque.rbegin() + numElements, results.begin());
-            }
-
-            return results;
+            return m_priorityQueue.size();
         }
 
         void clear()
@@ -288,27 +224,10 @@ class ThreadSafeDeque
             /* Aquire the lock */
             std::unique_lock<std::mutex> lock(m_mutex);
 
-            m_deque.clear();
-        }
-        
-        /* Recommended to implement a memoryUsage() method for your type, if not using
-           a simple type, and pass to the below function instead.
-           Otherwise, sizeof() is unlikely to be accurate. */
-        size_t memoryUsage() const
-        {
-            return m_deque.size() * sizeof(T) + sizeof(m_deque);
-        }
-
-        size_t memoryUsage(std::function<size_t(T)> memUsage) const
-        {
-            return std::accumulate(
-                m_deque.begin(),
-                m_deque.end(),
-                sizeof(m_deque),
-                [&memUsage](const auto acc, const auto item) {
-
-                return acc + memUsage(item);
-            });
+            while (!m_priorityQueue.empty())
+            {
+                m_priorityQueue.pop();
+            }
         }
 
     private:
@@ -335,7 +254,7 @@ class ThreadSafeDeque
                     return true;
                 }
 
-                return !m_deque.empty();
+                return !m_priorityQueue.empty();
             });
 
             /* Stopping, don't return data */
@@ -345,12 +264,12 @@ class ThreadSafeDeque
             }
 
             /* Get the first item in the queue */
-            item = m_deque.front();
+            item = m_priorityQueue.top();
 
             /* Remove the first item from the queue */
             if (removeFromQueue)
             {
-                m_deque.pop_front();
+                m_priorityQueue.pop();
             }
 
             /* Unlock the mutex before notifying, so it doesn't block after
@@ -363,9 +282,8 @@ class ThreadSafeDeque
             return item;
         }
         
-    private:
         /* The deque data structure */
-        std::deque<T> m_deque;
+        std::priority_queue<T, std::vector<T>, Comparison> m_priorityQueue;
 
         /* The mutex, to ensure we have atomic access to the queue */
         mutable std::mutex m_mutex;