--- a/src/WalletApi/ApiDispatcher.h
+++ b/src/WalletApi/ApiDispatcher.h
@@ -54,7 +54,8 @@ class ApiDispatcher
             const uint16_t bindPort,
             const std::string rpcBindIp,
             const std::string rpcPassword,
-            std::string corsHeader);
+            std::string corsHeader,
+            unsigned int walletSyncThreads = std::thread::hardware_concurrency());
 
         /////////////////////////////
         /* Public member functions */
@@ -378,4 +379,7 @@ class ApiDispatcher
 
         /* Used along with our password with pbkdf2 */
         CryptoPP::byte m_salt[16];
+
+        /* Amount of threads to use during wallet syncing */
+        unsigned int m_walletSyncThreads;
 };