--- a/src/WalletApi/ParseArguments.cpp
+++ b/src/WalletApi/ParseArguments.cpp
@@ -12,6 +12,8 @@
 #include <config/CryptoNoteConfig.h>
 #include <config/WalletConfig.h>
 
+#include <thread>
+
 #include "version.h"
 
 Config parseArguments(int argc, char **argv)
@@ -24,11 +26,14 @@ Config parseArguments(int argc, char **argv)
 
     int logLevel;
 
+    unsigned int threads;
+
     options.add_options("Core")
         ("h,help", "Display this help message", cxxopts::value<bool>(help)->implicit_value("true"))
 
         ("v,version", "Output software version information", cxxopts::value<bool>(version)->default_value("false")->implicit_value("true"))
-        ("log-level", "Specify log level", cxxopts::value<int>(logLevel)->default_value(std::to_string(config.logLevel)), "#");
+        ("log-level", "Specify log level", cxxopts::value<int>(logLevel)->default_value(std::to_string(config.logLevel)), "#")
+        ("threads", "Specify number of wallet sync threads", cxxopts::value<unsigned int>(threads)->default_value(std::to_string(std::max(1u, std::thread::hardware_concurrency()))), "#");
 
     options.add_options("Network")
         ("p,port", "The port to listen on for http requests",
@@ -81,6 +86,16 @@ Config parseArguments(int argc, char **argv)
     else
     {
         config.logLevel = static_cast<Logger::LogLevel>(logLevel);
+    }
+
+    if (threads == 0)
+    {
+        std::cout << "Thread count must be at least 1" << std::endl;
+        exit(1);
+    }
+    else
+    {
+        config.threads = threads;
     }
 
     return config;