--- a/src/WalletApi/WalletApi.cpp
+++ b/src/WalletApi/WalletApi.cpp
@@ -41,7 +41,7 @@ int main(int argc, char **argv)
         /* Init the API */
         api = std::make_shared<ApiDispatcher>(
             config.port, config.rpcBindIp, config.rpcPassword,
-            config.corsHeader
+            config.corsHeader, config.threads
         );
 
         /* Launch the API */