… src/WalletBackend/WalletBackend.h
… src/WalletBackend/WalletSynchronizer.cpp
… src/WalletBackend/WalletSynchronizer.h
… src/zedwallet++/Open.cpp
… src/zedwallet++/ParseArguments.cpp
… src/zedwallet++/ParseArguments.h