@zpalmtree 2019-05-19
The build now passes. I believe I have also fixed the crashes.