@zpalmtree 2019-05-19
Working on performance numbers now for a few different configurations and devices...